requirements.txt
openai
pillow
python-dotenv
pyexiv2